requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.0,<4.0
drf-yasg>=1.20.0